requirements.txt
djangorestframework
markdown
django-filter
uritemplate